+++ rtl/sys_gen/w11a/cmoda7/sys_w11a_c7.vhd
@@ -1,4 +1,4 @@
--- $Id: sys_w11a_c7.vhd 1342 2023-01-02 15:18:19Z mueller $
+-- $Id: sys_w11a_c7.vhd 1349 2023-01-11 14:52:42Z mueller $
 -- SPDX-License-Identifier: GPL-3.0-or-later
 -- Copyright 2017-2023 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
 --
@@ -28,6 +28,7 @@
 --
 -- Synthesized:
 -- Date         Rev  viv    Target       flop  lutl  lutm  bram  slic
+-- 2023-01-11  1349 2022.1  xc7a35t-1    3451  6019   279  50.0  2006
 -- 2023-01-02  1342 2022.1  xc7a35t-1    3434  6005   279  50.0  1969
 -- 2022-12-31  1340 2022.1  xc7a35t-1    3450  6018   279  50.0  1986
 -- 2022-12-27  1339 2022.1  xc7a35t-1    3454  6026   279  50.0  2013